Fixed AudioTimeout (hopefully); Refactored; Updated Dependencies;
