+++ src/index.ts
@@ -1,3 +1,2 @@
 export {default as AudioManager} from './audioManager';
-export type {VoiceConnectionDataModel} from './types/VoiceConnectionDataModel';
-export type {VoiceAudioDataModel} from './types/VoiceAudioDataModel';
+export type {VoiceConnectionDataModel, VoiceAudioDataModel} from './typings';
