+++ src/typings.d.ts
@@ -17,3 +17,16 @@ export type VoiceConnectionDataModel = {
      */
     VoiceAdapter: DiscordGatewayAdapterCreator;
 }
+
+export type VoiceAudioDataModel = {
+    /**
+     * The unique identifier for the audio resource.
+     */
+    ResourceType: "Link" | "File";
+
+    /**
+     * The URL or file path of the audio resource.
+     * any is to be assumed to require(filepath)
+     */
+    Resource: string | any;
+}
